feat: 添加产品调拨功能及优化相关枚举
