+++ router/api.go
@@ -98,7 +98,8 @@ func Api(g *gin.Engine) {
 				products.POST("/info", product.ProductController{}.Info)    // 产品详情
 				products.PUT("/update", product.ProductController{}.Update) // 产品更新
 
-				products.PUT("/damage", product.ProductController{}.Damage) // 产品报损
+				products.PUT("/damage", product.ProductController{}.Damage)      // 产品报损
+				products.POST("/allocate", product.ProductController{}.Allocate) // 产品调拨
 			}
 		}
 	}
