+++ router/init.go
@@ -8,7 +8,7 @@ import (
 
 func Init(r *gin.Engine) {
 	// 正则参数验证
-	r.Use(middlewares.RegexValidator())
+	r.Use(middlewares.CustomValidator())
 
 	Base(r)
 	Api(r)
