+++ types/enum.go
@@ -1,0 +1,6 @@
+package types
+
+type Enums interface {
+	InMap() error
+	ToMap() any
+}
